Split facade by service
