+++ ScoutSuite/providers/aws/services/awslambda/service.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from ScoutSuite.providers.base.configs.resources import Resources
 from ScoutSuite.providers.aws.configs.regions_config import Regions
-from ScoutSuite.providers.aws.facade import AWSFacade
+from ScoutSuite.providers.aws.facade.facade import AWSFacade
 from opinel.utils.aws import build_region_list
 
 
@@ -26,7 +26,7 @@ class RegionalLambdas(Resources):
         facade = AWSFacade()
 
         functions = {}
-        for raw_function in facade.get_lambda_functions(region):
+        for raw_function in facade.awslambda.get_functions(region):
             name, function = self.parse_function(raw_function)
             functions[name] = function
