+++ ScoutSuite/providers/aws/services/ec2/ami.py
@@ -1,5 +1,5 @@
 from ScoutSuite.providers.aws.configs.regions_config import ScopedResources
-from ScoutSuite.providers.aws.facade import AWSFacade
+from ScoutSuite.providers.aws.facade.facade import AWSFacade
 
 class AmazonMachineImages(ScopedResources):
     def __init__(self, owner_id):
@@ -7,7 +7,7 @@ class AmazonMachineImages(ScopedResources):
         self.facade = AWSFacade()
     
     async def get_resources_in_scope(self, region): 
-        return self.facade.get_ec2_images(region, self.owner_id)
+        return self.facade.ec2.get_images(region, self.owner_id)
 
     def parse_resource(self, raw_image):
         raw_image['id'] = raw_image['ImageId']
